Use instance type for constructor return type
I did this in 2 places:

1. Eagerly, in resolveAnonymousTypeMembers, which might not be feasible
in the end since it may lead to circularities when I start supporting
prototype assignments.
2. "Elsewhere", in getJSClassType, which is still called in a smattering
of places, nearly all Very Dangerous workarounds for circularity
problems themselves.

With (1), it is easy to make a copy of call signatures to construct
signatures, altering the return type to be the instance type of the
constructor function.